Enable tests to work on a non-Internet-connected host (#20)
* Enable tests to work on a non-Internet-connected host

We now populate Ivy's cache by running a simple PySpark-based
pytest unit test. One of the transitive-dependencies has a
range-based constraint on the joda-time library that requires
Ivy to search for the dependency, so we populate a local Maven
repository with just that library. We then tell Ivy to only
perform the search against that local repo.

* Update documentation

* Adjust test to make Sonarcloud happy